split storage db upgrade test
